Safari-611.1.10/LayoutTests/editing/selection/ios/do-not-hide-selection-in-visible-field-expected.txt


Click me
This test verifies platform selection UI is not incorrectly suppressed when the editable element is visible. To manually run the test, tap the button to focus the search field, and confirm that the caret view appears in the field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caretRect.top is 7
PASS caretRect.left is 11
PASS caretRect.width is 2
PASS caretRect.height is 25
PASS successfullyParsed is true

TEST COMPLETE
